Homework 6

+++ src/services/group.service.ts
@@ -5,7 +5,7 @@ import { getConnection, getManager } from 'typeorm';
 import { Group } from '../models/group.model';
 import { UserRepository } from '../repository/user.repositoty';
 import { HTTPError } from '../common/error.class';
-import { HttpCode } from '../util/const';
+import { HttpCode } from '../constants/http-code';
 
 export class GroupService {
     groups: IGroup[];
@@ -41,6 +41,12 @@ export class GroupService {
     }
 
     async delete(id: string) {
+        const existedGroup = await this.groupRepository.getOne(id);
+
+        if (!existedGroup) {
+            throw new HTTPError(HttpCode.NOT_FOUND, 'Group is not found');
+        }
+
         return await this.groupRepository.delete(id);
     }
